Update README.md
In this Project, I use Spring Boot to rapidly create a REST API. JPA and Hibernate are employed to manage the database interaction, with MySQL (or any RDB) as the backend database. The configuration is done entirely with annotations, eliminating the need for XML. Spring Boot’s built-in integration with JPA/Hibernate simplifies database operations and enhances developer productivity.